+++ test/utils/input/number-inputs.ts
@@ -49,23 +49,48 @@ export const nonFiniteNumberInputs: number[] = [
     -Infinity
 ];
 
-export const positiveNumberInputs: number[] = [
-    10,
-    10.01,
+export const zeroInputs: number[] = [0, -0];
+
+export const negativeFloatInputs: number[] = [
+    -Number.MIN_VALUE,
+    -Number.EPSILON,
+    -10.01,
+    -0.01,
+    -(1.0 / 3)
+];
+
+export const positiveFloatInputs: number[] = [
     Number.MIN_VALUE,
+    Number.EPSILON,
+    10.01,
+    0.01,
+    (1.0 / 3)
+];
+
+export const floatInputs: number[] = [
+    ...negativeFloatInputs,
+    ...positiveFloatInputs
+];
+
+export const negativeIntegerInputs: number[] = [
+    -Number.MAX_VALUE,
+    -Number.MAX_SAFE_INTEGER,
+    Number.MIN_SAFE_INTEGER,
+    -1,
+    -10,
+    -100,
+    -1000,
+    -1000.0
+];
+
+export const positiveNumberInputs: number[] = [
+    ...positiveFloatInputs,
+    10,
     Number.MAX_VALUE,
-    Number.MAX_SAFE_INTEGER,
-    Number.EPSILON
+    Number.MAX_SAFE_INTEGER
 ];
 
 export const negativeNumberInputs: number[] = [
-    -10,
-    -10.01,
-    -Number.MIN_VALUE,
-    -Number.MAX_VALUE,
-    -Number.MAX_SAFE_INTEGER,
-    -Number.EPSILON,
-    Number.MIN_SAFE_INTEGER
+    ...negativeFloatInputs,
+    ...negativeIntegerInputs
 ];
-
-export const zeroInputs: number[] = [0, -0];