+++ docs/releases.md
@@ -3,7 +3,7 @@ layout: post
 author:
   - Brittni Watkins
 date: 2026-05-18
-modified_date: 2026-05-20
+modified_date: 2026-06-16
 title: "Releases"
 tags:
   - releases
@@ -11,4 +11,5 @@ toc: true
 ---
 
 ## v0.x
-- [v0.1.0-alpha.0](releases/v0.x/v0.1.x/v0.1.0-alpha.x/v0.1.0-alpha.0/doc/index.html) *(Released May 19, 2026)*
+- [v0.1.0-alpha.0](releases/v0.x/v0.1.x/v0.1.0-alpha.x/v0.1.0-alpha.0/doc/index.html)
+- [v0.1.0-alpha.1](releases/v0.x/v0.1.x/v0.1.0-alpha.x/v0.1.0-alpha.1/doc/index.html)